# web/tooling/openapi/workforce.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Workforce.Host.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5041/"
    }
  ],
  "paths": {
...
    "/api/v1/recruitment/requisitions": {
      "get": {
        "tags": [
          "RecruitmentRequisitions"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/RequisitionStatus"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedRecruitmentResultOfJobRequisition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedRecruitmentResultOfJobRequisition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedRecruitmentResultOfJobRequisition"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RecruitmentRequisitions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequisitionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequisitionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequisitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/requisitions/{id}": {
      "get": {
        "tags": [
          "RecruitmentRequisitions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RecruitmentRequisitions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequisitionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequisitionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequisitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/requisitions/{id}/submit-approval": {
      "post": {
        "tags": [
          "RecruitmentRequisitions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/requisitions/{id}/approve": {
      "post": {
        "tags": [
          "RecruitmentRequisitions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/requisitions/{id}/open": {
      "post": {
        "tags": [
          "RecruitmentRequisitions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/requisitions/{id}/hold": {
      "post": {
        "tags": [
          "RecruitmentRequisitions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/requisitions/{id}/close": {
      "post": {
        "tags": [
          "RecruitmentRequisitions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/requisitions/{id}/cancel": {
      "post": {
        "tags": [
          "RecruitmentRequisitions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRequisition"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConcurrencyActionRequest": {
        "required": [
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "rowVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      },
...
      "CreateRequisitionRequest": {
        "required": [
          "organizationUnitId",
          "positionId",
          "locationId",
          "hiringManagerId",
          "recruiterId",
          "requisitionNumber",
          "titleEn",
          "titleAr",
          "openingsCount",
          "employmentType",
          "pipelineId",
          "requisitionReason",
          "targetStartDate"
        ],
        "type": "object",
        "properties": {
          "organizationUnitId": {
            "type": "string",
            "format": "uuid"
          },
          "positionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "locationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "hiringManagerId": {
            "type": "string",
            "format": "uuid"
          },
          "recruiterId": {
            "type": "string",
            "format": "uuid"
          },
          "requisitionNumber": {
            "type": "string"
          },
          "titleEn": {
            "type": "string"
          },
          "titleAr": {
            "type": "string"
          },
          "openingsCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "employmentType": {
            "type": "string"
          },
          "pipelineId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "requisitionReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetStartDate": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "JobRequisition": {
        "required": [
          "id",
          "tenantId",
          "legalEntityId",
          "organizationUnitId",
          "positionId",
          "locationId",
          "hiringManagerId",
          "recruiterId",
          "requisitionNumber",
          "titleEn",
          "titleAr",
          "openingsCount",
          "employmentType",
          "pipelineId",
          "pipelineVersion"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "$ref": "#/components/schemas/TenantId"
          },
          "legalEntityId": {
            "$ref": "#/components/schemas/LegalEntityId"
          },
          "organizationUnitId": {
            "type": "string",
            "format": "uuid"
          },
          "positionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "locationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "hiringManagerId": {
            "type": "string",
            "format": "uuid"
          },
          "recruiterId": {
            "type": "string",
            "format": "uuid"
          },
          "requisitionNumber": {
            "type": "string"
          },
          "titleEn": {
            "type": "string"
          },
          "titleAr": {
            "type": "string"
          },
          "openingsCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "employmentType": {
            "type": "string"
          },
          "pipelineId": {
            "type": "string",
            "format": "uuid"
          },
          "pipelineVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/RequisitionStatus"
          },
          "approvalRequestId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "requisitionReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetStartDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "openedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "closedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "rowVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      },
...
      "PagedRecruitmentResultOfJobRequisition": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobRequisition"
            }
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "RequisitionStatus": {
        "type": "integer"
      },
...
      "TenantId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}